+++ firmware/include/checksum.h
@@ -10,6 +10,6 @@
  * @param len 即 packet 的长度，单位是字节
  * @return 校验和无误则返回 true ，有误则返回 false
  */
-bool validateAndFillChecksum(uint8_t *packet, uint32_t len);
+bool validateAndFillChecksum(uint8_t *packet, uint16_t len);
 
 #endif
