+++ firmware/router.c
@@ -28,9 +28,9 @@ void icmp_error_gen() {
     dma_send_request();
 
     // 将原始包保留至 ICMPv6 错误包的数据部分
-    uint32_t len = (DMA_LEN + 48) > 1280 ? 1280 : (DMA_LEN + 48);
+    uint16_t len = (DMA_LEN + 48) > 1280 ? 1280 : (DMA_LEN + 48);
     DMA_LEN = len;
-    for (uint32_t i = len - 1; i >= 62; --i) {
+    for (uint16_t i = len - 1; i >= 62; --i) {
         DMA_PTR[i] = DMA_PTR[i - 48];
     }
